pyproject.toml
[project]
name = "sih25"
version = "0.1.0" 
description = "FloatChat - AI-Powered Conversational Interface for ARGO Ocean Data Discovery and Visualization"
requires-python = ">=3.10"
dependencies = [
    # Core Data Processing & Scientific Computing
    "xarray>=2024.1.0",
    "netcdf4>=1.6.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "pyarrow>=15.0.0",
    "cftime>=1.6.0",
    # Workflow Orchestration
    "prefect>=2.18.0",
    # Data Validation & Schema
    "pydantic>=2.5.0",
    # Database & Storage
    "supabase>=2.3.0",
    "psycopg2-binary>=2.9.0",
    "sqlalchemy>=2.0.0",
    # Vector Database & Embeddings
    "chromadb>=0.4.0",
    "sentence-transformers>=2.2.0",
    # Web Framework & API
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    # Visualization & Dashboard
    "plotly>=5.17.0",
    "dash>=2.14.0",
    "streamlit>=1.28.0",
    # Model Context Protocol & LLM Integration
    "openai>=1.3.0",
    "anthropic>=0.7.0",
    "fastapi-mcp>=0.4.0",
    # Voice Processing (Pipecat) - Using more compatible versions
    "speechrecognition>=3.10.0",
    "pyaudio>=0.2.11",
    # Geospatial Processing
    "geopandas>=0.14.0",
    "folium>=0.15.0",
    # Development & Testing
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.9.0",
    "isort>=5.12.0",
    "flake8>=6.1.0",
    # Configuration & Environment
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.1",
    "click>=8.1.0",
    # HTTP & Async Processing
    "httpx>=0.25.0",
    "aiofiles>=23.2.0",
    "asyncio-mqtt>=0.16.0",
    "matplotlib>=3.9.4",
    "seaborn>=0.13.2",
    "nbconvert>=7.16.6",
    "agno>=2.0.7",
    "groq>=0.31.1",
    # Voice AI Integration Dependencies
    "loguru>=0.7.0",
    "pipecat-ai[openai,deepgram]>=0.0.30",
    "aiohttp>=3.8.0",
]

[build-system]
requires = ["hatchling>=1.18.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["sih25"]

[tool.uv]
package = true
